type errors

+++ Project1/Emitter.cpp
@@ -35,8 +35,8 @@ CEmitter::CEmitter(CPokeOrbitApp * pokeOrbit)
 	mPokemonTime = 0;
 	mPokeStopTime = 0;
 	mt19937_64 rand(chrono::system_clock::now().time_since_epoch().count());
-	mTimeOfNextPokemonEmission= rand() % 12 + 3;
-	mTimeOfNextPokeStopEmission = rand() % 10 + 20;
+	mTimeOfNextPokemonEmission= (double)(rand() % 12 + 3);
+	mTimeOfNextPokeStopEmission = (double)(rand() % 10 + 20);
 }
 
 /// Destructor
@@ -91,13 +91,13 @@ void CEmitter::Update(double elapsed)
 	mPokeStopTime += elapsed;
 	if (mPokemonTime >= mTimeOfNextPokemonEmission)
 	{
-		mTimeOfNextPokemonEmission =  rand() % 12 + 3;
+		mTimeOfNextPokemonEmission = (double)(rand() % 12 + 3);
 		mPokemonTime = 0;
 		EmitPokemon();
 	}
 	if (mPokeStopTime >= mTimeOfNextPokeStopEmission)
 	{
-		mTimeOfNextPokeStopEmission = rand() % 10 + 20;
+		mTimeOfNextPokeStopEmission = (double)(rand() % 10 + 20);
 		mPokeStopTime = 0;
 		EmitPokeStop();
 	}
